End Lesson 01.06.2016
This is the state of the projet at the end of the lesson of the first june 2016. This version have a major upgrade in design and fonctionnalities. After-Tommorrow,  I will change some little bug and add some features. I nearly can started with the setup of the message system for the forum.

Orion, Project Leader